use serde_multikey in DefaultDocMapper
