+++ quickwit/quickwit-doc-mapper/src/doc_mapper.rs
@@ -235,7 +235,7 @@ mod tests {
     use crate::default_doc_mapper::{FieldMappingType, QuickwitJsonOptions, TextIndexingOptions};
     use crate::{
         Cardinality, DefaultDocMapper, DefaultDocMapperBuilder, DocMapper, DocParsingError,
-        FieldMappingEntry, ModeType, TermRange, WarmupInfo, DYNAMIC_FIELD_NAME,
+        FieldMappingEntry, Mode, TermRange, WarmupInfo, DYNAMIC_FIELD_NAME,
     };
 
     const JSON_DEFAULT_DOC_MAPPER: &str = r#"
@@ -366,7 +366,7 @@ mod tests {
     #[test]
     fn test_doc_mapper_query_with_json_field_default_search_fields() {
         let doc_mapper: DefaultDocMapper = DefaultDocMapperBuilder {
-            mode: ModeType::Dynamic,
+            mode: Mode::default(),
             ..Default::default()
         }
         .try_build()
@@ -385,7 +385,7 @@ mod tests {
     #[test]
     fn test_doc_mapper_query_with_json_field_ambiguous_term() {
         let doc_mapper: DefaultDocMapper = DefaultDocMapperBuilder {
-            mode: ModeType::Dynamic,
+            mode: Mode::default(),
             ..Default::default()
         }
         .try_build()
